Fix front-end startup
The front-end uses a Nuxt-based application (as a default) built
using yarn. The original Dockerfile started it with
`ENTRYPOINT ["yarn"]` and `CMD ["nuxt-start"]. This did *not* work.

This was changed to `ENTRYPOINT ["yarn"]` and `CMD ["run start"],
but that did *not* work, either.

Finally, the original was changed back as it turned out that
the issue was the building itself. Now first enter the
`frontend` directory and run `npm install`, then run the containers.

```shell
cd frontend
npm install
cd ..
docker-compose up -d
```